Cached and versioning
It was dded an fuction to get new updates in the app asking the user if they would like to update it to refresh it. Aditional some information was cahed to avoid loading the data again like the font and the icons. Some changes were made in the view in order to test the functionallity of the new versioning function.