Project restructuring and updated
The whole project was restructured.
The shared components are more organized and understandable using modules for each folder. It was better categorized and links were updated to redirect to the right components. Additional the designs were updated the name structure for a simpler using only the component name instead of the design name.