Material added, Signup set up and some models updated
Angular Material was added to the project but is currently not being used. Maybe in the future will help us with anything that might show up. The signup form was worked in the design and also in the redirection for the component and allow to share data between the routes. Additional the user model was updated.