Logo and brand images, fonts, palette and organization were applied
The logo and the brand images were change for a more modern and better looking one and added to the svgs files and change in some components like the navigation.

The font was changed from Bahnschrift Light to Barlow and will be use across the platform along with Segoe UI.

The palette and variables were changed however this might change in the future due to a dark and light mode inside the platform. The signin and signup components were updated for this changes but still needs some more work on them.

Some organization was also done in the styles' folder separating the files to make it more readable.

Additional the lp-banner image was change for a better one.