User model re-named, Auth generated and signin working
Some of the user model properties were re-named in order to have a more consist naming with the fireauth API. The Auth service is currently working and should not have big changes further from now. Additional was also implemented in the sign in page and currently is working. Is missing the sign up to be created to implement it in that page as well.