Re-structured part 2 xD
Since the past commit 69510da940a5c888808c95c3f6ed3ba594e53d2c did not clicked it was changed once again. This time for good since the way to continue the programing the project will change with a state manager. It was also installed ngxs per previous said.

In this commit the app is not working correctly as before and the ngxs implementation in the next commits will fix everything.