Authentication module update to use NGXS and user model change
The authentication module was change to use the new structure and as mention before in the commit 91c1fb26ec8e25b78bd12a786a3468cb77e8e538 now NGXS is being use to control the state of the form. Is not completed the authentication module but now is working as previously.

Additional the user model was change to contain a funtion to return an empty user and the userform since there's fields missing needed for the form.