Settings module updated to work as before
The module was chnaged and now the component does not create a component dynamically but uses the routes in order to navigate across the different childrens routes it has. The navigation was also update for a better responsive and is some how working good.

Additional some changes were made in the controls module in the style that are not used.