This fixes #3 and implements #2 a little bit
The issue for #3 was easily adding the property overflow auto in the scss file.

However, the #2 implementation of separating the menu and the navigation was done and is working but is quite extensive both in scss as in inputs properties in order to work in different components. This implementation most be reduced in code in the future but as for now should be working as before.

Additional the menu was added to the navigation and the setting but the remaining components could not be deleted since the menu could not be implemented as a navigation bar. Therefore, further changes must be done in the component and it should completely replace the components in the navigation.