Basic design updated for fonts and fonts added in ngsw
The basic design was updated for the change were the fonts were changed (8181d0fc46144d1517d598c3e277668af703307a).  Due to this change it was looking quite weird and cluster the text. This was updated in order to look better and as well the navigation was change to look better. Still is in progress but is looking way better than before.

Additional the fonts were removed and now will be added in the cache with the PWA ngsw.