Authentication added, guards and navigation and other components were changed a little
The authentication component was added and this is going to be the container for the different child components, since the routes are directly from the root the way it was implemented is that each component will have as itself as a child.

The guard for a signed and unsigned user was added allowing to access some routes and not some other if you're signed in or not.

Some minor changed were done as well in other components like the navigation and the footer in to update them.

Finally, the folder for svgs was deleted and their content was switch to the imgs folder that also was renamed as images for more simplicity.