This implements a little of #5
For performance related issues the data was store in the local and session storage but using NGXS state won't be secure. This not implements a way to store some information no sensitive keeping the performance and removing some information. However, this still needs more work since when a user gets called will return all the data. It needs to return only the data the user set as public.