This fixes #4 and fixes other issues discovered while fixing the component
This issue was fixed by creating a class open-modal in the body that can be use later from a modal. It only adds the overflow as hidden and that prevents the scroll movement when the menu is open.

Additional another problem was found when you click the outside of the open menu to close it will trigger the element behind the shadow. This was also fixed by making the shadow a div and checking for the click in that element to close it. This also was implemented in the pop menu. When you're in the pop menu the scroll won't trigger the close of the button and this is the correct functioning for the component, might change in the future but for now it works great.