Pricing page, container width and some little fixes
The pricing component was worked to show a little taste of how the page will look later, still is missing a lot since the table needs to show the feature price individual. But I thinks is looking good so far.

The container class was updated to have a smaller max-width to look better.

And some fixes with the medias were done as well in the authentication module.