Icon, header and banner image updated
The icons for the PWA app were updated for the need ones. The header was also change, however, the current design does not look as good as I was expecting, this will be change in the future for another header and along that change the footer needs to be updated to stay in the bottom if there's not enogh content without relying in the size of other components.

Finally the pages were updated as firebase functions.