Changes in the basic design
The basic design was updated to not have the navigation since will be replace in the future with the navigation from the controls but it needs to be implemented a floating navigation and also was removed the messages along with all its child components, this because a future zone will be added as "dashboard" where the users will be able to see their chats and other additional information.

Other changes added in the design were some little changes in order to reduce code and responsiveness.

Other modules were updated a little but it was only to remove the imports not used and add the enviroment in the state manager.