Authentication service along with storage class update
The authentication was changed to cache the logged user in the local storage, for this the class Storage was created, it handles the local and session storage as well as it convert it to an observable to listen to the changes. These changes are only listen if the change is done from the app and not by the user accessing to the dev tools of the browser and is intended to work this way to avoid changing information and see reflected in the app.

Additional the other components depending on the auth service were updated to work with this change.