Storage class renamed and implemented in the profile handler
The storage classes created in the past commit (6c122d577671fbc40355afd4187115083c9d86a1) was renamed to StorageLS since it was confused with the Storage class from the system. Additional it was implemented in the profile handler to cache the user loaded.

This not works on the #5 issue but loads better the page with lesser code.

Some minor issues in the routes were also updated.