Models and db structure with some updates.
The firebase project was split into two different projects, one for development environment and the other one for production. This was updated in the Angular project as well with some scripts in the node package to handle the deploy and building of the app in each environment.

The rules for firestore and storage will be handled from now on in the firebase console and therefore removed from the project.

Since when the project started the firestore db wasn't completely structured some issues were starting to happening and the whole app had to be changed since a model had to be updated to handle new information some how. Due to this was determined that all the models in the db have to be made to avoid this situation. Most models were created and tested to work fine. Some models still need to be finished like the one for user that even if is already finished it needs to be updated in the modules since it change quite a bit. Other models like the one for payment and product are still being worked since it needs slidejs and I still don't know what output would I be able to save in the db. Until a get there I won't be able to do it.

Some of the tsconfig configuration were changed for a smaller bundle after building the app.

Some other minnor changes were also done in the README.