Post in the basic design worked
The cards in the posts page was already built show how would it look in the basic design. Individual post still is being work on and some animation in the cards still needs more work to be finished but the layout looks close to the final product.

Other fixes were done like the container class, since using cover-ui has its own class no need for the one in the main styles file. Also, was change some padding and margins for other components like the toolbar and the cover.

Additional some scripts were added to have an easier environment developing and deploy the app.