Markdown support completed
The MarkdownComponent was deleted since it was change to a pipe and the output will be inserted in a div dynamically. The support is almost finished since is a string from Firestore still needs to process the sanitized content Firebase does to clean the code. This will be work on in the mdToHtml pipe. The styling was also added in the styles being universal to the app was wrapped in a div to avoid conflicts with other classes. Also, the basic design was updated to support this new change.